Type pmag as float to match H5DLP::Particle members
H5DLP::Particle::{px,py,pz,ke,mass} are all float. Declaring pmag
as double caused a type mismatch in downstream calculations that
assume float, so make it float to match.

+++ src/EDepSimPersistencyManager.cc
@@ -600,7 +600,8 @@ void EDepSim::PersistencyManager::SummarizeTrajectoriesH5(
         part.px = ndTraj->GetInitialMomentum().x();
         part.py = ndTraj->GetInitialMomentum().y();
         part.pz = ndTraj->GetInitialMomentum().z();
-        double pmag = std::sqrt(pow(part.px,2) + pow(part.py,2) + pow(part.pz,2));
+
+        float pmag = std::sqrt(pow(part.px,2) + pow(part.py,2) + pow(part.pz,2));
         part.ke = std::sqrt(pow(pmag,2) + pow(part.mass,2)) - part.mass;
 
         // Get the first point
